+++ beatmap/v2/difficulty.ts
@@ -10,6 +10,8 @@ import { SpecialEventsKeywordFilters } from './specialEventsKeywordFilters.ts';
 import { deepCopy } from '../../utils/misc.ts';
 
 export class DifficultyData extends Serializable<IDifficultyData> {
+    private _fileName = 'UnnamedDifficulty.dat';
+
     version: `2.${0 | 2 | 4 | 5 | 6}.0`;
     notes: Note[];
     sliders: Slider[];
@@ -26,9 +28,7 @@ export class DifficultyData extends Serializable<IDifficultyData> {
         this.obstacles = data._obstacles.map((obj) => Obstacle.create(obj));
         this.events = data._events.map((obj) => Event.create(obj));
         this.waypoints = data._waypoints.map((obj) => Waypoint.create(obj));
-        this.specialEventsKeywordFilters = SpecialEventsKeywordFilters.create(
-            data._specialEventsKeywordFilters,
-        );
+        this.specialEventsKeywordFilters = SpecialEventsKeywordFilters.create(data._specialEventsKeywordFilters);
         this.customData = data._customData;
     }
 
@@ -58,6 +58,17 @@ export class DifficultyData extends Serializable<IDifficultyData> {
             _specialEventsKeywordFilters: this.specialEventsKeywordFilters?.toObject(),
             _customData: deepCopy(this.customData),
         };
+    }
+
+    set fileName(name: string) {
+        this._fileName = name.trim();
+    }
+    get fileName() {
+        return this._fileName;
+    }
+    setFileName(fileName: string) {
+        this.fileName = fileName;
+        return this;
     }
 
     getFirstInteractiveTime = (): number => {
@@ -91,10 +102,7 @@ export class DifficultyData extends Serializable<IDifficultyData> {
         let obstacleEnd = 0;
         for (let i = this.obstacles.length - 1; i >= 0; i--) {
             if (this.obstacles[i].isInteractive()) {
-                obstacleEnd = Math.max(
-                    obstacleEnd,
-                    this.obstacles[i].time + this.obstacles[i].duration,
-                );
+                obstacleEnd = Math.max(obstacleEnd, this.obstacles[i].time + this.obstacles[i].duration);
             }
         }
         return obstacleEnd;