+++ converter/V2toV3.ts
@@ -20,25 +20,16 @@ const tag = (name: string) => {
  * ---
  * **WARNING:** Custom data may be lost on conversion, as well as other incompatible attributes.
  */
-export const V2toV3 = (
-    data: DifficultyDataV2,
-    skipPrompt?: boolean,
-): DifficultyDataV3 => {
+export const V2toV3 = (data: DifficultyDataV2, skipPrompt?: boolean): DifficultyDataV3 => {
     if (!skipPrompt) {
-        logger.warn(
-            tag('V2toV3'),
-            'Converting beatmap v2 to v3 may lose certain data!',
-        );
+        logger.warn(tag('V2toV3'), 'Converting beatmap v2 to v3 may lose certain data!');
         const confirmation = prompt('Proceed with conversion? (y/N):', 'n');
         if (confirmation![0].toLowerCase() !== 'y') {
             throw Error('Conversion to beatmap v3 denied.');
         }
         logger.info(tag('V2toV3'), 'Converting beatmap v2 to v3');
     } else {
-        logger.warn(
-            tag('V2toV3'),
-            'Converting beatmap v2 to v3 may lose certain data!',
-        );
+        logger.warn(tag('V2toV3'), 'Converting beatmap v2 to v3 may lose certain data!');
     }
     const template = v3.DifficultyData.create();
 
@@ -78,16 +69,10 @@ export const V2toV3 = (
                 };
             }
             if (typeof n.customData._fake === 'boolean') {
-                logger.warn(
-                    tag('V2toV3'),
-                    `notes[${i}] at time ${n.time} NE _fake will be removed.`,
-                );
+                logger.warn(tag('V2toV3'), `notes[${i}] at time ${n.time} NE _fake will be removed.`);
             }
             if (typeof n.customData._cutDirection === 'number') {
-                logger.debug(
-                    tag('V2toV3'),
-                    `notes[${i}] at time ${n.time} NE _cutDirection will be converted.`,
-                );
+                logger.debug(tag('V2toV3'), `notes[${i}] at time ${n.time} NE _cutDirection will be converted.`);
             }
         }
         if (n.isBomb()) {
@@ -116,8 +101,7 @@ export const V2toV3 = (
                     c: n.type as 0 | 1,
                     x: n.lineIndex,
                     y: n.lineLayer,
-                    d: n.cutDirection >= 1000 ||
-                            typeof n.customData?._cutDirection === 'number'
+                    d: n.cutDirection >= 1000 || typeof n.customData?._cutDirection === 'number'
                         ? n.cutDirection === 8 ? 8 : 1
                         : clamp(n.cutDirection, 0, 8),
                     a: a,
@@ -158,10 +142,7 @@ export const V2toV3 = (
                 };
             }
             if (typeof o.customData._fake === 'boolean') {
-                logger.warn(
-                    tag('V2toV3'),
-                    `obstacles[${i}] at time ${o.time} NE _fake will be removed.`,
-                );
+                logger.warn(tag('V2toV3'), `obstacles[${i}] at time ${o.time} NE _fake will be removed.`);
             }
         }
         template.obstacles.push(
@@ -215,10 +196,7 @@ export const V2toV3 = (
                         lerpType: e.customData._lerpType,
                     };
                     if (e.customData._propID) {
-                        logger.warn(
-                            tag('V2toV3'),
-                            `events[${i}] at time ${e.time} Chroma _propID will be removed.`,
-                        );
+                        logger.warn(tag('V2toV3'), `events[${i}] at time ${e.time} Chroma _propID will be removed.`);
                     }
                     if (e.customData._lightGradient) {
                         logger.warn(
@@ -237,10 +215,7 @@ export const V2toV3 = (
                         direction: e.customData._direction,
                     };
                     if (e.customData._reset) {
-                        logger.warn(
-                            tag('V2toV3'),
-                            `events[${i}] at time ${e.time} Chroma _reset will be removed.`,
-                        );
+                        logger.warn(tag('V2toV3'), `events[${i}] at time ${e.time} Chroma _reset will be removed.`);
                     }
                     if (e.customData._counterSpin) {
                         logger.warn(
@@ -248,15 +223,8 @@ export const V2toV3 = (
                             `events[${i}] at time ${e.time} Chroma _counterSpin will be removed.`,
                         );
                     }
-                    if (
-                        e.customData._stepMult ||
-                        e.customData._propMult ||
-                        e.customData._speedMult
-                    ) {
-                        logger.warn(
-                            tag('V2toV3'),
-                            `events[${i}] at time ${e.time} Chroma _mult will be removed.`,
-                        );
+                    if (e.customData._stepMult || e.customData._propMult || e.customData._speedMult) {
+                        logger.warn(tag('V2toV3'), `events[${i}] at time ${e.time} Chroma _mult will be removed.`);
                     }
                 }
                 if (e.isLaserRotationEvent()) {
@@ -405,11 +373,18 @@ export const V2toV3 = (
                         scale: e._scale,
                         position: e._position?.map((n) => n * 0.6) as Vector3,
                         rotation: e._rotation,
-                        localPosition: e._localPosition?.map(
-                            (n) => n * 0.6,
-                        ) as Vector3,
+                        localPosition: e._localPosition?.map((n) => n * 0.6) as Vector3,
                         localRotation: e._localRotation,
                         lightID: e._lightID,
+                    };
+                }) ?? [];
+                continue;
+            }
+            if (k === '_pointDefinitions') {
+                template.customData.pointDefinitions = data.customData._pointDefinitions!.map((e) => {
+                    return {
+                        name: e._name,
+                        points: e._points,
                     };
                 }) ?? [];
                 continue;