+++ example/printInfo.ts
@@ -1,4 +1,4 @@
-import * as bsmap from '../mod.ts';
+import * as bsmap from 'https://deno.land/x/bsmap/mod.ts';
 
 export const printChromaEnvironment = (d: bsmap.v3.DifficultyData) => {
     const envEnh = d.customData.environment ?? [];
@@ -26,10 +26,7 @@ export const printV3Event = (d: bsmap.v3.DifficultyData) => {
         '\nLight Color Event Box',
         d.lightColorEventBoxGroups.reduce((t, e) => t + e.events.length, 0),
         '\nLight Color',
-        d.lightColorEventBoxGroups.reduce(
-            (t, e) => t + e.events.reduce((r, y) => r + y.events.length, 0),
-            0
-        )
+        d.lightColorEventBoxGroups.reduce((t, e) => t + e.events.reduce((r, y) => r + y.events.length, 0), 0)
     );
     console.log();
     console.log(
@@ -38,20 +35,11 @@ export const printV3Event = (d: bsmap.v3.DifficultyData) => {
         '\nLight Rotation Event Box',
         d.lightRotationEventBoxGroups.reduce((t, e) => t + e.events.length, 0),
         '\nLight Rotation',
-        d.lightRotationEventBoxGroups.reduce(
-            (t, e) => t + e.events.reduce((r, y) => r + y.events.length, 0),
-            0
-        )
+        d.lightRotationEventBoxGroups.reduce((t, e) => t + e.events.reduce((r, y) => r + y.events.length, 0), 0)
     );
     console.log(
         '\nTotal V3 Event',
-        d.lightColorEventBoxGroups.reduce(
-            (t, e) => t + e.events.reduce((r, y) => r + y.events.length, 0),
-            0
-        ) +
-            d.lightRotationEventBoxGroups.reduce(
-                (t, e) => t + e.events.reduce((r, y) => r + y.events.length, 0),
-                0
-            )
+        d.lightColorEventBoxGroups.reduce((t, e) => t + e.events.reduce((r, y) => r + y.events.length, 0), 0) +
+            d.lightRotationEventBoxGroups.reduce((t, e) => t + e.events.reduce((r, y) => r + y.events.length, 0), 0)
     );
 };